# frontend/travel-with-anna-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification for Travel with Anna API",
    "description": "OpenApi documentation for Travel with Anna API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/twa/v1",
      "description": "Local server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/avatar": {
      "get": {
        "tags": [
          "Avatar"
        ],
        "operationId": "getCurrentUserAvatar",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Avatar"
        ],
        "operationId": "uploadAvatar",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/admin": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "operationId": "updateUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserAdminUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
...
    "/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "operationId": "getAllUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseUserAdminResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/user/{identifier}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "operationId": "getUserByIdentifier",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAdminResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/avatar/{userId}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "operationId": "getAvatar",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarImg"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/{userId}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "operationId": "deleteUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserAdminDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "UserAdminEditRequest": {
        "required": [
          "roleName",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "accountLocked": {
            "type": "boolean"
          },
          "enabled": {
            "type": "boolean"
          },
          "roleName": {
            "type": "string"
          }
        }
      },
      "UserAdminUpdateRequest": {
        "required": [
          "userAdminEditRequest"
        ],
        "type": "object",
        "properties": {
          "userAdminEditRequest": {
            "$ref": "#/components/schemas/UserAdminEditRequest"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "PageResponseUserAdminResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserAdminResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "UserAdminResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "userName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "avatar": {
            "type": "string",
            "format": "byte"
          },
          "accountLocked": {
            "type": "boolean"
          },
          "enabled": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date"
          },
          "roleName": {
            "type": "string"
          }
        }
      },
      "AvatarImg": {
        "type": "object",
        "properties": {
          "avatar": {
            "type": "string",
            "format": "byte"
          }
        }
      },
...
      "UserAdminDeleteRequest": {
        "required": [
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "password": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}